scheduler(test) Test 2 and int64_t in scheduler
Added a test for the scheduler (kill and exit)
Added int64_t in the scheduler and in the program definitions (or else cast
from pointers to int will give errors)

+++ src/kernel/context_init.c
@@ -19,7 +19,7 @@ static void kernel_run(int (*runf) (void *), void arg, unsigned int magic)
 
 
 void context_kernelinit(
-		context_t *c, kernel_stack_t stack, int (*runf) (void *), void *arg)
+		context_t *c, kernel_stack_t stack, int64_t (*runf) (void *), void *arg)
 {
 	c->satp = get_kernel_satp().reg;
 	c->ra = (void*) crt_process;
